Moved group-import underneath the groups folder
